fix: Implement C++ stdout forwarding to Kit server console
- Add stdout capture mechanism using select() and non-blocking I/O
- Forward C++ application stdout in real-time to Kit server using same WebSocket format as build messages
- Fix binary path resolution to handle CMake/Makefile builds
- Optimize stdout responsiveness with 300ms monitoring cycles
- Add comprehensive logging with fire emoji markers for debugging
- Fix C++ compilation error with type casting

This fixes the core issue where CMake build output appeared in Kit server but C++ cout statements did not.

+++ kuksa-syncer/app/main.cpp
@@ -75,7 +75,7 @@ public:
         
         // Light levels (natural + artificial)
         std::normal_distribution<float> light_noise(0.0f, 20.0f);
-        float natural_light = 500.0f * std::max(0.0f, std::sin(time_of_day * M_PI));
+        float natural_light = 500.0f * std::max(0.0f, (float)std::sin(time_of_day * M_PI));
         float artificial_light = occupied_hours ? 200.0f : 50.0f;
         light_lux = natural_light + artificial_light + light_noise(gen);
